Consolidate circle approx in dedicated module
